+++ Thunderstore/CHANGELOG.md
@@ -5,11 +5,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [3.3.0] - 2025-10-21 - "Vanilla models"
+## [3.3.0] - 2025-10-22 - "Vanilla models"
 ### Added
 - Added the vanilla meshes of Poster5 and Tips.
   - There's an option in the `Settings` section to use one of them or neither of them.
 - Support for transparent image backgrounds has been added. Instead of transparent backgrounds appearing black, they are now fully transparent.
+- Added support for multiple poster packs within a single mod folder.
+  - Backward compatibility maintained for existing poster packs.
+- Added multi-pack support.
+  - Multiple packs inside 1 mod can be detected.
 
 ### Changed
 - Poster positions and rotation of ShipWindows, TwoStoryShip, and WiderShipMod were improved to be closer to the original.